change balance data type in sql to numeric

+++ migrations/000001_init-schema.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "users" (
   "id" varchar PRIMARY KEY,
-  "balance" bigint NOT NULL,
+  "balance" NUMERIC NOT NULL,
   "currency" varchar NOT NULL DEFAULT 'RUB',
   "created_at" timestamptz NOT NULL DEFAULT (now())
 );
@@ -9,7 +9,7 @@ CREATE TABLE "deposits" (
   "id" bigserial PRIMARY KEY,
   "from_user_id" varchar DEFAULT 'atm machine',
   "to_user_id" varchar NOT NULL,
-  "amount" bigint NOT NULL,
+  "amount" NUMERIC NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT (now()),
   "comment" varchar DEFAULT 'deposit'
 );
@@ -18,7 +18,7 @@ CREATE TABLE "withdrawals" (
   "id" bigserial PRIMARY KEY,
   "from_user_id" varchar NOT NULL,
   "to_user_id" varchar DEFAULT 'atm machine',
-  "amount" bigint NOT NULL,
+  "amount" NUMERIC NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT (now()),
   "comment" varchar DEFAULT 'withdrawal'
 );
@@ -27,7 +27,7 @@ CREATE TABLE "transactions" (
   "id" bigserial PRIMARY KEY,
   "from_user_id" varchar NOT NULL,
   "to_user_id" varchar NOT NULL,
-  "amount" bigint NOT NULL,
+  "amount" NUMERIC NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT (now()),
   "comment" varchar DEFAULT 'money transfer'
 );
